+++ crates/biome_cli/src/commands/lint.rs
@@ -5,6 +5,7 @@ use crate::commands::{
 use crate::{
     execute_mode, setup_cli_subscriber, CliDiagnostic, CliSession, Execution, TraversalMode,
 };
+use biome_configuration::linter::RuleCode;
 use biome_configuration::vcs::PartialVcsConfiguration;
 use biome_configuration::{
     PartialConfiguration, PartialFilesConfiguration, PartialLinterConfiguration,
@@ -24,6 +25,7 @@ pub(crate) struct LintCommandPayload {
     pub(crate) vcs_configuration: Option<PartialVcsConfiguration>,
     pub(crate) files_configuration: Option<PartialFilesConfiguration>,
     pub(crate) paths: Vec<OsString>,
+    pub(crate) rule: Option<RuleCode>,
     pub(crate) stdin_file_path: Option<String>,
     pub(crate) staged: bool,
     pub(crate) changed: bool,
@@ -38,6 +40,7 @@ pub(crate) fn lint(session: CliSession, payload: LintCommandPayload) -> Result<(
         cli_options,
         mut linter_configuration,
         mut paths,
+        rule,
         stdin_file_path,
         vcs_configuration,
         files_configuration,
@@ -128,6 +131,7 @@ pub(crate) fn lint(session: CliSession, payload: LintCommandPayload) -> Result<(
         Execution::new(TraversalMode::Lint {
             fix_file_mode,
             stdin,
+            rule,
         })
         .set_report(&cli_options),
         session,