+++ crates/biome_cli/src/execute/std_in.rs
@@ -156,11 +156,17 @@ pub(crate) fn run<'a>(
             }
         }
 
+        let rule = if let TraversalMode::Lint { rule, .. } = mode.traversal_mode() {
+            *rule
+        } else {
+            None
+        };
         if !mode.is_check_apply_unsafe() {
             let result = workspace.pull_diagnostics(PullDiagnosticsParams {
                 categories: RuleCategories::LINT | RuleCategories::SYNTAX,
                 path: biome_path.clone(),
                 max_diagnostics: mode.max_diagnostics.into(),
+                rule,
             })?;
             diagnostics.extend(result.diagnostics);
         }